requirements-colab.txt
accelerate>=0.34.0
bitsandbytes>=0.43.0
datasets>=2.21.0
evaluate>=0.4.3
jinja2>=3.1.4
matplotlib>=3.9.2
peft>=0.12.0
pydantic>=2.8.2
PyMuPDF>=1.24.9
python-docx>=1.1.2
reportlab>=4.2.2
soundfile>=0.12.1
openai-whisper>=20231117
sentencepiece>=0.2.0
transformers>=4.44.2
trl>=0.10.1
scikit-learn>=1.5.1
pyyaml>=6.0.2
mediapipe>=0.10.14
numpy>=1.26.4
